+++ firecore/torch/__init__.py
@@ -5,3 +5,8 @@ try:
     import torch
 except ImportError:
     logger.exception("please install pytorch first")
+finally:
+    from . import cudnn_utils
+    from . import jit_utils
+    from . import module_utils
+    from . import optimizer_utils