+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `config.from_file` and `config.from_snippet`
 - Export `config`
 - Export `resolve`
+- `firecore.torch.cudnn_utils.setup_cudnn_benchmark()`
 
 ### Changed